server changes

+++ RMIStoreServer/src/rmistore/commons/interfaces/Account.java
@@ -3,9 +3,8 @@ package rmistore.commons.interfaces;
 import java.rmi.Remote;
 import java.rmi.RemoteException;
 import rmistore.commons.exceptions.Rejected;
-
 public interface Account extends Remote {
-    public float getBalance() throws RemoteException;
+    public double getBalance() throws RemoteException;
 
     public void deposit(float value) throws RemoteException, Rejected;
